+++ src/iotsaBLEServer.h
@@ -69,6 +69,7 @@ public:
   using IotsaBLEServerModBaseMod::IotsaBLEServerModBaseMod;
   void setup() override;
   void serverSetup() override;
+  void lateSetupDone() override;
   void loop() override;
 #ifdef IOTSA_WITH_WEB
   String info() override;